refactor(brain): split memory.ts into filter/recall/query + rename fileStoreRouter (A17.4)
brain/memory.ts (993 lines) mixed three unrelated engines with memory CRUD:
  - filter.ts (120)  — the recall FilterExpression DSL: validate, lower to a
                       Mongo filter, lower to a native $vectorSearch prefilter
  - recall.ts (516)  — the recall engine: recall/recallGlobal/findSimilar/
                       checkDuplicates + the doc -> RecallResult mapping
  - query.ts  (125)  — structured queryBrain: operator whitelist, ReDoS-safe
                       sanitiser, projection guard that never emits `embedding`
  - memory.ts (267)  — remember + memory CRUD

Dependencies run strictly one way (memory -> recall -> filter; filter/query
import no siblings), so there are no cycles: remember reaches into recall.ts
only for the optional insert-time duplicate check. Importers repointed:
api/brain/search.ts, brain/entities.ts, brain/dupe-scanner.ts, the MCP
file/search/shared tools, and two tests importing compiled queryBrain /
mergeEmbeddingExclusion.

Also in this PR (both requested):
- Renamed filesRouter -> fileStoreRouter. It sat beside the brain's
  fileMetaRouter and the pair read as one API; they are bytes-on-disk vs the
  brain record describing a file. Store/Meta makes it visible at the call site.
- Docs: the integration guide described what document conversion produces but
  never WHEN. Since MCP write_file went async (A10), an agent that writes a doc
  and immediately recalls it finds nothing. Added a Timing section (what each
  surface returns, how to poll embeddingStatus, media + "text" bypass cases).
- Disabled MD060 in .markdownlint.json: it flagged 416 lines because it
  disagrees with the repo's deliberate compact table style used in every table.
  Logged the remaining 318 pre-existing docs lint errors (172 are code fences
  with no language) as A18 rather than leave them unrecorded.

Verified: build clean, no dead imports (tsc --noUnusedLocals); brain 183/183,
mcp-tools 94/94, proxy-spaces 44/44, schema 29+85, standalone 800/0,
sync 191/0, red-team 272/0.
