+++ .markdownlint.json
@@ -1,4 +1,5 @@
 {
   "MD013": false,
-  "MD024": { "siblings_only": true }
+  "MD024": { "siblings_only": true },
+  "MD060": false
 }
