+++ server/src/api/brain/search.ts
@@ -7,7 +7,10 @@ import { Router } from 'express';
 import { requireSpaceAuth, denyReadOnly } from '../../auth/middleware.js';
 import { globalRateLimit } from '../../rate-limit/middleware.js';
 import { NotFoundError } from '../../util/errors.js';
-import { countMemories, queryBrain, findSimilar, recall, validateFilterExpression, type RecallKnowledgeType, type FilterExpression } from '../../brain/memory.js';
+import { countMemories } from '../../brain/memory.js';
+import { queryBrain } from '../../brain/query.js';
+import { findSimilar, recall, type RecallKnowledgeType } from '../../brain/recall.js';
+import { validateFilterExpression, type FilterExpression } from '../../brain/filter.js';
 import { traverseGraph, traverseRecallSeeds, MAX_RECALL_TRAVERSE, resolveEdgeEntityNames } from '../../brain/edges.js';
 import { memoryEmbedText, entityEmbedText, edgeEmbedText, chronoEmbedText, fileEmbedText } from '../../brain/embed-text.js';
 import { embed } from '../../brain/embedding.js';
